Change behaviour of getAuthority
The TrayContentProvider#mAuthority will be never empty.
It will set automatically by the Android system when the ContentProvider got created.
The ContentProvider lives as long as the procress run. So it can't be never null.

Whatever. If it happen we have implemented tray in a wrong way.
So I've removed the log and added a RuntimeExcpetion.